+++ src/c_lib/auth/packets.hpp
@@ -14,7 +14,7 @@ class auth_token_CtoS: public FixedSizeReliableNetPacketToServer<auth_token_CtoS
 
         char token[AUTH_TOKEN_LENGTH+1];
 
-    inline void packet(char* buff, unsigned int* buff_n, bool pack)
+    inline void packet(char* buff, size_t* buff_n, bool pack)
     {
         pack_string(this->token, AUTH_TOKEN_LENGTH+1, buff, buff_n, pack);
     }
@@ -26,7 +26,7 @@ class auth_token_valid_StoC: public FixedSizeReliableNetPacketToClient<auth_toke
 {
     public:
 
-    inline void packet(char* buff, unsigned int* buff_n, bool pack)
+    inline void packet(char* buff, size_t* buff_n, bool pack)
     {
     }
 
@@ -37,7 +37,7 @@ class auth_token_invalid_StoC: public FixedSizeReliableNetPacketToClient<auth_to
 {
     public:
 
-    inline void packet(char* buff, unsigned int* buff_n, bool pack)
+    inline void packet(char* buff, size_t* buff_n, bool pack)
     {
     }
 
@@ -49,7 +49,7 @@ class clock_time_StoC: public FixedSizeReliableNetPacketToClient<clock_time_StoC
     public:
         uint64_t server_time;
 
-    inline void packet(char* buff, unsigned int* buff_n, bool pack)
+    inline void packet(char* buff, size_t* buff_n, bool pack)
     {
         pack_u64(&server_time, buff, buff_n, pack);
     }
