+++ src/c_lib/item/toolbelt/net/CtoS.hpp
@@ -8,7 +8,7 @@ class toolbelt_set_slot_CtoS: public FixedSizeReliableNetPacketToServer<toolbelt
     public:
         uint8_t slot;
 
-    inline void packet(char* buff, unsigned int* buff_n, bool pack)
+    inline void packet(char* buff, size_t* buff_n, bool pack)
     {
         pack_u8(&slot, buff, buff_n, pack);
     }
@@ -18,7 +18,7 @@ class toolbelt_set_slot_CtoS: public FixedSizeReliableNetPacketToServer<toolbelt
 class toolbelt_begin_alpha_action_CtoS: public FixedSizeReliableNetPacketToServer<toolbelt_begin_alpha_action_CtoS>
 {
     public:
-    inline void packet(char* buff, unsigned int* buff_n, bool pack)
+    inline void packet(char* buff, size_t* buff_n, bool pack)
     {
     }
     inline void handle();
@@ -27,7 +27,7 @@ class toolbelt_begin_alpha_action_CtoS: public FixedSizeReliableNetPacketToServe
 class toolbelt_end_alpha_action_CtoS: public FixedSizeReliableNetPacketToServer<toolbelt_end_alpha_action_CtoS>
 {
     public:
-    inline void packet(char* buff, unsigned int* buff_n, bool pack)
+    inline void packet(char* buff, size_t* buff_n, bool pack)
     {
     }
     inline void handle();
@@ -36,7 +36,7 @@ class toolbelt_end_alpha_action_CtoS: public FixedSizeReliableNetPacketToServer<
 class toolbelt_beta_action_CtoS: public FixedSizeReliableNetPacketToServer<toolbelt_beta_action_CtoS>
 {
     public:
-    inline void packet(char* buff, unsigned int* buff_n, bool pack)
+    inline void packet(char* buff, size_t* buff_n, bool pack)
     {
     }
     inline void handle();
